refactor out temporary minimal api endpoint
