+++ Starlights.sln
@@ -58,8 +58,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "components", "components", 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "aspire", "aspire", "{27E01094-52D2-4E80-B911-2B6CE218F09A}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "components", "components", "{8C6408C7-ACE5-4543-BF21-0C7EECFF8509}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "systems", "systems", "{8BED2340-C2DE-4C9B-8CBC-E2B09165C292}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Starlights.Platform.Components.FastEndpoints", "src\platform\components\Starlights.Platform.Components.FastEndpoints\Starlights.Platform.Components.FastEndpoints.csproj", "{554E31B4-0DAB-47D1-BB00-C669FD63C16E}"
@@ -154,12 +152,11 @@ Global
 		{2072DE9B-881E-413F-8B61-2403BF4A1405} = {6ED99F10-9BC8-4FE3-B34C-AB17B8F00D61}
 		{150E417D-A00E-400C-A1EB-C66DA23BC16E} = {6ED99F10-9BC8-4FE3-B34C-AB17B8F00D61}
 		{F9A7BDB2-9DCF-4E16-9934-6BE0B83CC284} = {D42B1339-D2AE-42CB-9AF5-78612A3DFB10}
-		{49D2699A-427E-4388-B9DF-3C6E5EBF0146} = {8C6408C7-ACE5-4543-BF21-0C7EECFF8509}
+		{49D2699A-427E-4388-B9DF-3C6E5EBF0146} = {6ED99F10-9BC8-4FE3-B34C-AB17B8F00D61}
 		{BF7DB35B-E014-4B73-A491-627122312C7C} = {C840D2F9-B821-4C84-B2D2-5CB18C1C9F23}
 		{50AD13D4-A4B1-423E-88D5-9985BA9A87B7} = {6ED99F10-9BC8-4FE3-B34C-AB17B8F00D61}
 		{16CB40A1-26BB-4950-AB2E-37F37798D00A} = {128F28F2-7131-4C7B-A386-8E6984195241}
 		{D42B1339-D2AE-42CB-9AF5-78612A3DFB10} = {C840D2F9-B821-4C84-B2D2-5CB18C1C9F23}
-		{8C6408C7-ACE5-4543-BF21-0C7EECFF8509} = {6ED99F10-9BC8-4FE3-B34C-AB17B8F00D61}
 		{8BED2340-C2DE-4C9B-8CBC-E2B09165C292} = {4CA9D0EC-9739-4751-B08B-DF40599F2A52}
 		{554E31B4-0DAB-47D1-BB00-C669FD63C16E} = {D42B1339-D2AE-42CB-9AF5-78612A3DFB10}
 		{0ECC6BED-E1AB-4DBB-A4A3-9B72B1C2BE57} = {6ED99F10-9BC8-4FE3-B34C-AB17B8F00D61}
