+++ src/apps/Starlights.Application/appsettings.Development.json
@@ -6,6 +6,6 @@
     }
   },
   "ConnectionStrings": {
-    "DefaultConnection": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=starlights"
+    "DefaultConnectionEx": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=starlights"
   }
 }
