+++ src/modules/elements/Modules.Elements.Endpoints/Modules.Elements.Endpoints.csproj
@@ -1,11 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
+	<PropertyGroup>
+		<RootNamespace>Starlights.Modules.Elements.Endpoints</RootNamespace>
+	</PropertyGroup>
+	
 	<ItemGroup>
 		<PackageReference Include="FastEndpoints" />
 	</ItemGroup>
